[WIP] Trying to spin auth into its own service
