+++ apps/dashboard/dashboard.js
@@ -1,48 +1,10 @@
 'use strict';
 const path = require('path');
+
 const applicationService = require('../../services/applications');
+const auth = require('../../services/auth');
 
-module.exports = function (app) {
+module.exports = function (router) {
 
-    app.route('/:formId?/:applicationId?').get(async (req, res, next) => {
-        try {
-            const {formId, applicationId} = req.params;
 
-            const viewData = {
-                formId,
-                forms: [],
-                applications: [],
-                formTitle: null,
-                applicationData: null
-            };
-
-            if (applicationId) {
-                // look up a specific application
-                viewData.applicationData = await applicationService.getApplicationsById(applicationId);
-
-                if (!viewData.applicationData) {
-                    return next();
-                }
-
-                viewData.formTitle = viewData.applicationData.formTitle;
-            } else if (formId) {
-                // get applications for a given form
-                viewData.applications = await applicationService.getApplicationsByForm(formId);
-
-                if (viewData.applications.length < 1) {
-                    return next();
-                }
-
-                viewData.formTitle = viewData.applications[0].formTitle;
-            } else {
-                // fetch all forms instead for a list
-                viewData.formTitle = 'All online forms';
-                viewData.forms = await applicationService.getAvailableForms();
-            }
-
-            res.render(path.resolve(__dirname, 'views/dashboard'), viewData);
-        } catch (error) {
-            next(error);
-        }
-    });
 };
