+++ services/bodyParser.js
@@ -1,0 +1,4 @@
+'use strict';
+const bodyParser = require('body-parser');
+
+module.exports = [bodyParser.json(), bodyParser.urlencoded({ extended: false })];
